bpo-45229: Make tkinter tests discoverable
